abstract out populateFramesAndColumns
