+++ src/modelInput/modelInput.js
@@ -1,6 +1,7 @@
 import { createControl } from './createControl';
 import { createGridableValues } from './createGridableValues';
 import { h2oModelBuilderForm } from './h2oModelBuilderForm/h2oModelBuilderForm';
+import { populateFramesAndColumns } from './populateFramesAndColumns';
 
 import { requestModelBuilders } from '../h2oProxy/requestModelBuilders';
 
@@ -22,58 +23,6 @@ export function modelInput() {
       return false;
     });
     const _modelForm = Flow.Dataflow.signal(null);
-    const populateFramesAndColumns = (frameKey, algorithm, parameters, go) => {
-      const destinationKeyParameter = lodash.find(parameters, parameter => parameter.name === 'model_id');
-      if (destinationKeyParameter && !destinationKeyParameter.actual_value) {
-        destinationKeyParameter.actual_value = `${algorithm}-${Flow.Util.uuid()}`;
-      }
-
-      //
-      // Force classification.
-      //
-      const classificationParameter = lodash.find(parameters, parameter => parameter.name === 'do_classification');
-      if (classificationParameter) {
-        classificationParameter.actual_value = true;
-      }
-      return _.requestFrames(_, (error, frames) => {
-        let frame;
-        let frameKeys;
-        let frameParameters;
-        let parameter;
-        let _i;
-        let _len;
-        if (error) {
-          // empty
-          // TODO handle properly
-        } else {
-          frameKeys = (() => {
-            let _i;
-            let _len;
-            const _results = [];
-            for (_i = 0, _len = frames.length; _i < _len; _i++) {
-              frame = frames[_i];
-              _results.push(frame.frame_id.name);
-            }
-            return _results;
-          })();
-          frameParameters = lodash.filter(parameters, parameter => parameter.type === 'Key<Frame>');
-          for (_i = 0, _len = frameParameters.length; _i < _len; _i++) {
-            parameter = frameParameters[_i];
-            parameter.values = frameKeys;
-
-            // TODO HACK
-            if (parameter.name === 'training_frame') {
-              if (frameKey) {
-                parameter.actual_value = frameKey;
-              } else {
-                frameKey = parameter.actual_value;
-              }
-            }
-          }
-          return go();
-        }
-      });
-    };
     ((() => requestModelBuilders(_, (error, modelBuilders) => {
       _algorithms(modelBuilders);
       _algorithm(_algo ? lodash.find(modelBuilders, builder => builder.algo === _algo) : void 0);
@@ -84,7 +33,7 @@ export function modelInput() {
         if (builder) {
           algorithm = builder.algo;
           parameters = flowPrelude.deepClone(builder.parameters);
-          return populateFramesAndColumns(frameKey, algorithm, parameters, () => _modelForm(h2oModelBuilderForm(_, algorithm, parameters)));
+          return populateFramesAndColumns(_, frameKey, algorithm, parameters, () => _modelForm(h2oModelBuilderForm(_, algorithm, parameters)));
         }
         return _modelForm(null);
       });
